Extract etcd operations to their own file
